feat(audio): move AudioMonitor class to a separate file
